swasthya_ai_backend/requirements.txt
fastapi
uvicorn
google-generativeai
pinecone
python-dotenv
